Removed VideoValidator.
VideoValidator was using youtube_dl, which was kind of a pain to have as a dependency. It meant people had to have ffmpeg, and calling other processes isn't very elegant either. Furthermore, the validator was only being used for YouTube URL validation, and we can do this easily with a few API calls. Therefore, youtube_dl is no longer a dependency (isodate is now one however).

We should see some performance gains overall, potentially significant by not having to wait around for YouTubeDL/FFMPEG.

The goal for the future is to try and use existing APIs to parse duration info from the most popular sites posted on the subreddit.